Fix/feedback pagination

+++ packages/demo-feedback-form/src/components/Comment.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useContext } from "react";
-import { Box, Button, ListItem } from "@alexmcgovern/boondoggle.design";
+import { Box, Button, Card } from "@alexmcgovern/boondoggle.design";
 import { FirebaseContext } from "@alexmcgovern/firebase";
 import { faTrashCan } from "@fortawesome/free-solid-svg-icons";
 import { deleteDoc } from "firebase/firestore";
@@ -7,29 +7,50 @@ import type { CommentShape } from "../types";
 import { StarsDisplay } from "./StarsDisplay";
 
 export function Comment({
-  displayName,
   email,
   created,
   description,
   rating,
   documentRef,
   author_uid,
+  // intentionally destructured & unused until display name added to firebase auth
+  // eslint-disable-next-line unused-imports/no-unused-vars
+  displayName,
+  ...rest
 }: CommentShape) {
   const { user } = useContext(FirebaseContext) || {};
 
   const handleDelete = useCallback(async () => {
-    return deleteDoc(documentRef).catch((error) => {
+    if (!documentRef) return;
+
+    deleteDoc(documentRef).catch((error) => {
       /** ToDo(feedback-form/Comment): Render potential errors nicely in frontend */
       console.error(error);
     });
   }, [documentRef]);
 
   return (
-    <ListItem
-      title={`${displayName} (${email}) says...`}
-      subtitle={new Date(created.seconds * 1000).toLocaleString()}
-      description={description}
-    >
+    <Card {...rest}>
+      <Box as="header">
+        <Box fontSize="body_lg" fontWeight="medium">
+          {email}
+        </Box>
+
+        {created && (
+          <Box
+            color="neutral_text_lowContrast"
+            fontSize="body_sm"
+            fontWeight="medium"
+          >
+            {new Date(created.seconds * 1000).toLocaleString()}
+          </Box>
+        )}
+      </Box>
+
+      <hr />
+
+      {description && <Box as="p">{description}</Box>}
+
       <hr />
 
       <Box display="flex" alignItems="center" justifyContent="space-between">
@@ -45,6 +66,6 @@ export function Comment({
           </Button>
         )}
       </Box>
-    </ListItem>
+    </Card>
   );
 }
